src/functions/requirements.txt
azure-functions-core-tools
azure-functions
azure-identity
azure-ai-textanalytics==5.3.0
pyodbc
python-dotenv
